Update README.md to reflect requirement to label pods with user name.
Updated README.me to reflect the upcoming enforcement of the Informatics policy that all resources established on the Informatics share of the EIDF GPU cluster must be labeled with "eidf/user: <eidf user name>". This allows us to keep track of who created/owns/is responsible for specific k8s artefacts.